requirements.txt
pymupdf
pytesseract
pandas
openpyxl
Pillow